separate admin js for resolving conflict
